style: [FFBV-58] RecommendCards UI 수정
